Rename fixture files for clarity
- CsvPlugin.js → csv-plugin.js: kebab-case matches real-world plugin
  naming conventions (consistent with README examples)
- ExpectedAscii.js → AsciiSnapshot.js: "snapshot" is the established
  term for expected-output fixtures; drops "Expected" which belongs in
  the assertion, not the fixture file
