+++ test/fixtures/AsciiSnapshot.js
