+++ test/tests/analysis/BasicAnalysisTest.js
@@ -5,7 +5,7 @@ import VarietyHarness from '../../helpers/VarietyHarness.js';
 const test = new VarietyHarness('test', 'users');
 
 import sampleData from '../../fixtures/SampleData.js';
-import expectedAscii  from '../../fixtures/ExpectedAscii.js';
+import expectedAscii from '../../fixtures/AsciiSnapshot.js';
 
 describe('Basic Analysis', () => {
